+++ packages/sqlite_async/lib/src/sqlite_queries.dart
@@ -137,4 +137,10 @@ mixin SqliteQueries implements SqliteWriteContext, SqliteConnection {
       return tx.executeBatch(sql, parameterSets);
     });
   }
+
+  /// Ensures that all connections are aware of the latest schema changes applied (if any).
+  /// Queries and watch calls can potentially use outdated schema information after a schema update.
+  Future<void> refreshSchema() {
+    return get("PRAGMA table_info('sqlite_master')");
+  }
 }